Another POC of doing variables without Meriyah
This time, no code analysis needed! In a nutshell, this means we can
keep the naive parsing (which is fast) and then just... _not_ parse the
resulting JavaScript function.

The way I've approached this is by creating some sort of "strict mode"
(related: https://github.com/ventojs/vento/issues/101). The variables
are not detected from the code itself, but rather by what's passed in
the `it` object. Since this can change, the template function _can_
re-compile itself if it finds new keys have been added to the `it`
object, but presumably this doesn't happen very often for people opting
in to a strict mode.

Only 8 of the tests are failing, I think most of which I reckon _should_
fail in a "strict" mode, so I think this is quite a promising strategy.