Improve benchmarking
The benchmarks were quite simple, to the point they weren't a very
accurate measure of how Vento compares to its peers. I've created a more
elaborate template, that includes some basic logic and printing. This'll
(hopefully) make compilation times and render times a bit more
reflective of what they are in the wild.

I've restructured the benchmarks a bit to make it easier to add
additional languages, and added JSX, Pug and EJS to the list.

There are now 3 benchmarks; one for compilation speed (this does not
render anything), one for only rendering (excluding compilation) and one
for the combination. "Edge" has been excluded from the compilation-only
and render-only benchmarks, because it doesn't seem to have a
straight-forward API that would let us do these things separately.

To add a benchmark, add a template within `bench/templates/` and write
a function in `bench/setup.ts` that returns a "render" function (i.e. a
function that receives data and spits out a string).

If you're curious, JSX wins the compilation benchmark, Vento wins the
render one, and Eta wins the combined one.