Genericize tokenzation of echo tags
The echo tag got some special handling within tokenization. Specifically
it was transformed to an inline echo; for example, `{{echo}}foo{{/echo}}`
was transformed to `{{echo "foo"}}`. This special handling isn't really
needed, creates the perception of unreliable tokenization, and
complicates making changes to the tokenization process.

It also created a subtle bug where it overwrote all of the whitespace
trimming markers, because the tag was overwritten without explicitly
handling their presence. This points to a deeper issue; custom plugins
(including the trim plugin) can no longer rely on working together with
the echo tag. Working with the `trim` plugin is a special case, since we
cannot reliably detect what tags are echo tags within the tokenization
process in general. However, plugins that transform tokens to include
capture-style echo tags could not work, as the echo plugin only handled
inline echo tags.

As a bonus, this simplifies the signature of the internal `tokenize`
function, since we throw the "Unclosed echo tag" error within the echo
plugin. This makes a bit more sense anyway; the `tokenize` function
ideally shouldn't throw even if tags are unclosed (it even allows an
unclosed `{{#`).