Rework `analyze` function signature
The original `analyze` function takes a `visitor` argument, and this is
a bit awkward to work with in isolation. To do performance
optimizations, it is more ergonomic for this function to be lower level,
taking only primitive arguments, reducing the entanglement with other
parts of the codebase.

As such, I've reworked it to a `topLevel` function that reads a string
of JavaScript until it encounters a character that "breaks" out of the
top-level context; specifically this happens when it encounters a
top-level `}`, `]`, or `|>`. The algorithm remains mostly the same,
since conceptually the `analyze` function and `topLevel` function are
doing very similar things.

The "include" tag here is a bit of an outlier. Its syntax is actually
rather complex to get right, e.g. `include resolve({path: 'foo.tmpl'})
{ bar: 'baz' })` is difficult to parse (both with `analyze` as well as
`topLevel`). In this rework I've assumed that the object passed as data
is passed literally, i.e. starts with `{` and ends with a matching `}`.
I think that's a reasonable requirement to avoid the ambiguity.