Performance improvements (removing Meriyah)
This is a redo of https://github.com/ventojs/vento/pull/123, without the
`js.ts` or `tokenizer.ts`-specific parts. It iterates the generated code
to detect used variables using the same `iterateTopLevel()` function,
and then prepends a bit of code to define said variables.

It does speed up Vento quite a bit, though the "render" phase (which was
insanely fast before) is a little less fast (although still beating
nearly every other templating language). The render phase is
insignificant compared to compilation, but if a template is reused a
lot, this can add up.