Setup for more readable errors
There are, as I see it, 4 types of errors that can be thrown in Vento:

- Vento syntax errors, e.g. an unclosed `{{`)
- Tag syntax errors, e.g. `{{ for , }}` (these originate from plugins)
- JS syntax errors, e.g. `{{ foo bar }}`
- Runtime errors, e.g. `{{ (1).toUpperCase() }}`

The four all happen in various different ways, but ideally are presented
to the user in a similar format. In this commit, I've gone for a format
that shows the last three lines up to the error, pinpoints the position
of the error in the last of said lines, and displays the error. An
example of a JavaScript syntax error:

```
SyntaxError: Unexpected identifier 'bar'
 3 |     Totally fine.
 4 |     No errors here.
 5 |     {{ foo bar }}
                ^ Unexpected identifier 'bar'

Original error:
```

Some colors are used to improve readability and recognizability.

Current known limitations:

- Tabs need to be "measured"; currently the `^` pinpointing the location
  of the error is incorrect if the line contains tabs before said
  position.
- Some errors are unclear, e.g. the error thrown in `{{ foo? }}` says
  "Unexpected token ')'" which is weird for the user since they did not
  type a `)` at all. I think this might be solvable by checking if the
  error position coincides with the end of the raw tag source (in this
  case, the raw tag source is `foo?`, transformed to `__exports.content
  += (foo?) ?? "";`).