Show off selling points on homepage
Some feedback from the Eleventy Discord; it was mentioned that, from the
homepage, it looks like Vento's main feature is the syntax change. This
is not really accurate, as Vento has many things going for it that go
beyond just being nicer to write. I've written what I think are the four
main selling points; this includes the syntax change, but is last on the
list.