Support empty and more complex function arguments
Previously, the `{{ function }}` tag did not support something like
`{{ function hello() }}` since there was nothing within the parentheses.
It probably makes more sense to allow it since functions with arguments
must use parentheses and so people might want to use parentheses for
consistency.

Additionally, arguments containing `)` anywhere were disallowed, but
this is not a necessary restriction since the regular expression forces
the arguments to be captured in their entirety by requiring the closing
`\)` to be matched next to the end `$` of the tag.