Add a `strict` mode
This mode, which only makes sense when `autoDataVarname` is `true`,
throws an error when an undeclared variable is used in a template.
Currently Vento does not throw when doing something like `{{ foo }}` no
matter whether not `foo` is declared. This can be problematic because
typos can go unnoticed.

See also https://github.com/ventojs/vento/issues/101