Add separate benchmark for strict mode
It uses a separate template, since the default Vento template references
a `user` variable that doesn't exist and this throws in strict mode.

Note that the strict mode outperforms non-strict Vento in compilation
and full benchmarks, because it skips iterating the template for
variable name collection. It does lose by a lot in terms of rendering
speed (I'm seeing a 10x slowdown) but non-strict Vento is insanely fast
to render so the benefit in compilation speed still outweighs the
slowdown in rendering for the compilation and full benchmarks.