Fix and simplify escaping of JSON (and other)
The loop when escaping starts at the first index of a character that
needs to be escaped. In the tests, this ends up always being the first
characters, since we're testing something that starts with `<h1>`. For
JSON it's generally the second character, `"`, that needs escaping, and
so the first character gets skipped.

I've also taken the liberty of removing the loop entirely and replacing
it with a much simpler `.replace()` call.