Make closing {{ /layout }} tag optional
In this approach I add an argument to the `compileTokens()` function, to
allow for an optional end token.

The alternative, adding an extra `{{ /layout }}` token to the `tokens`
array within the layout plugin, ends up being quite complex. This is
because if there are two unclosed `{{ layout }}` tags, then inserting
the implied `{{ /layout }}` token actually gets eaten by the second
layout tag, and the first layout tag then _still_ ends up throwing an
error about being unclosed. Additionally, we probably want to also
remove the inserted end token if it wasn't used. Altogether I felt that
this way of doing it would get needlessly complicated.