Strict mode is not (always) slow
Of course this is a bit minor, but I don't think it's accurate to say
it's "much slower". The performance profile is different, but in some
cases it actually ends up being _faster_ than Vento's default. Template
compilations specifically is faster than usual because the generated
JavaScript doesn't need to be iterated over to look for variable names.
It's much slower in rendering, because it has to re-compile a new JS
function, but rendering overall is already orders of magnitude faster
than compilation, so this slowdown actually doesn't make it slower
overall. For projects that have a lot of templates being rendered only
once, the strict mode may prove to be faster.

I tried strict mode in lume.land's website, and (with some minor
modifications to get around undefined variables) it ends up building in
about 5.8s on my machine (averagish over three runs) whereas non-strict
mode builds in about 5.9s. That's a very minor difference, but at least
shows that strict mode doesn't have to be slow.