Fix variable detection with spread operator
Templates that only include a variable used after a spread operator fail
to detect the variable since it is instead falsely detected as property
access (i.e. `.foo`). We should make sure that if we see `.foo` that it
is not `...foo`, since in the latter case we do need to include it.