Autotrimming improvements
Autotrimming was done asymmetrically across comments and tags;
specifically, comments trimmed a newline character on their left, while
tags trimmed one on their right. This can create edge cases in certain
scenarios.

Additionally, filters weren't handled properly. With tag autotrimming,
the next token was taken to trim whitespace from, but this is incorrect
if the next token is a filter. Instead, we need to skip the filters and
trim the next "string"-type token.

A few tests have been added as well to solidify this behavior.