Add missing break statement
Just noticed that there's a `break` statement missing here. This is not
problematic per se, since a string token should never be followed up
with another string token nor a filter, but this is not something the
code should rely on. Adding it here to avoid confusion or bugs in the
future.