Allow direct data for include and layout tags
Allows passing data as a variable, rather than an object literal, both
for the `{{ include }}` tag and `{{ layout }}` tag.

Since paths can include JavaScript themselves, as in, for example,
`{{ include "./template" + ext }}`, this requires a bit of care. This
solution decides whether the last "word" in the tag is a data argument
by looking at what is in front of it; if it is a quote character of any
kind, or a word character, the last name token is a data variable.

Naturally, there are edge cases. Currently this does not allow for
something like `data.foo` or `data[0]`, but presumably can be extended
to do so.