Add `empty` filter
I considered `isEmpty` (and `is_empty`) as well, but I went with the
simpler `empty` because its use is mostly limited to `{{ if }}`
statements. It's unlikely users are directly printing booleans, and the
filter is more-or-less useless in tags like `{{ echo }}` or `{{ for }}`.
Given the relatively narrow scope, I don't think there are readability
issues with omitting the `is` from the name. Generally it'll be used in
contexts where booleans are expected, like

```vto
{{ if items |> empty }}
  {{ set isMissingTags = tags |> empty }}
{{ /if }}
```